Add Arena to the game
Arena will eventually be the container component
for the board and the scoreboard

Currently Arena contains just the board
Hoisted the game state up into the arena
and added a current game context for the board
to fetch the game.

To facilitate updating the game, added a "mini redux"
implementation to select the game from the state
as well as provide a dispatcher and actions to
update that piece of state.