Added some telemetry
Added a fake telemetry api that logs the telemetry data into the console.

Added some helpers to reduce needed boilerplate for consuming the api.

Hooked up telemetry to capture
1) mount/unmount
of the Match Scoreboard based on change of the
current game id.
2) mount/unmount events of the Board.
3) actions reduced by the root reducer
4) call duration for dispatching the add game
5) call duration for dispatching placement of
tokens on the board

Also fixed issue with Square component where we were passing false
instead of undefined to the onClick prop.