Add win detection based on last move
Added a new function to the game logic to
detect the winner based on the last move.
With this logic, we can focus only on testing
the rows which involve the last placed token.

Modified the Board component to consume
this new function. The original function is still
consumed by the game scoreboard.